Break up ConfigModal into smaller components
