+++ src/core/components/controls/ConfigControl.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react';
 import BaseControl from './BaseControl';
 import Modal from 'react-modal';
 import modalStyle from '@data-story-org/core/src/utils/modalStyle';
-import ConfigModal from '../modals/ConfigModal';
+import ConfigModal from '../modals/Config';
 import { Store } from '../../store';
 
 interface Props {
