+++ src/core/components/modals/Config/index.ts
@@ -1,0 +1,3 @@
+import ConfigModal from './Config';
+
+export default ConfigModal;
